Coming Soon notification updates (#44)
* Update Launch Your Site banner

Updated the "Coming Soon" banner on the Bluehost homepage to have a red border, red flag and the title was changed to "Coming Soon Active". I also changed the Launch button text to "Launch your site"

* Added Coming Soon Active Admin Bar item

* Update coming-soon.php

* Fix panel-warning rules

There was a comma in the wrong place within this:
#mojo-wrapper.mojo-bluehost-branding .panel.panel-warning,
#mojo-wrapper.mojo-bluehost-india-branding .panel.panel-warning { 
  border-left: 3px solid #F89C24;
}
Also changed the orange color to match the orange used in other places

* Remove flag from Coming Soon Active banner

* Changed Coming Soon Active background color

Changed Coming Soon Active admin menu item background color to orange #F89C24

* Update the panel div with the panel-warning class

Removed the panel-error class and replaced with panel-warning in order to get the orange border.

* Add notice-warning to Coming Soon notice

This removes the .notice-info class and adds the .notice-warning class to the admin notice that shows when the Coming Soon page is activated on a non-Bluehost hosted site. Done for consistency of colors used in all cases where the Coming Soon page is activated.

Thanks @desrosj