Remove the `mm_buy_now()` function.
Previously, this sent a request to the Mojo API to create an order and used the response to determine where the user should be redirected to in order to check out. Because we are going to remove the PayPal and other checkout options entirely, we can remove this in favor of a URL based approach where the parameters will be defined as query arguments.