Play nice with the new auto-update setting in WP 5.6.
WP 5.6 adds a UI element that allows site owners to opt-in to auto-updates for major version releases, and new installs will be opted-in by default.

Because the plugin was using the `allow_major_auto_core_updates` filter to implement the auto-update behavior chosen in the plugin, the text on the Core updates screen in the admin was incorrect. The core setting could also be toggled, but it would not affect the behavior of updates because of the plugin's filter.

This changes the behavior of the plugin for auto-updates on sites using WordPress 5.6 or higher to "step out of the way" to let the new code in Core manage the updates.