Sync plugin and theme auto-update settings to default WordPress Core options.
The plugin includes an option to blanket enable or disable auto-updates for plugins and/or themes. When either option is disabled, the user is able to manage auto-update settings on an individual basis using the default interface in WordPress Core introduced in version 5.5.

This works great when the plugin is active, but if the plugin is deactivated or removed, plugin/theme auto-updates are also wholesale turned off and the corresponding settings in Bluerock have no affect (they currently update the option key specific to this plugin and not the default WordPress Core option).

This updates the logic in the plugin to update the default WordPress Core auto-update settings for plugins and themes any time:

- a plugin or theme is installed, deleted, or updated.
- the plugin or theme auto-update setting is toggled.

When either of those two settings are toggled off, the Core option is not updated. This ensures all plugins and themes continue to receive auto-updates unless the user explicitly disables them on the plugin or theme page.

Bluerock and the installer scripts will soon be updated to also sync the default WordPress Core options to ensure auto-updates are appropriately applied even when the Bluehost plugin is disabled or removed.