Plugin/theme auto-update improvements (#97)
* Remove duplicate `switch` statement.

Because the `true` and `default` conditions are the same, we can omit the one for `true` and allow the `default` to handle the scenario.

* Sync plugin and theme auto-update settings to default WordPress Core options.

The plugin includes an option to blanket enable or disable auto-updates for plugins and/or themes. When either option is disabled, the user is able to manage auto-update settings on an individual basis using the default interface in WordPress Core introduced in version 5.5.

This works great when the plugin is active, but if the plugin is deactivated or removed, plugin/theme auto-updates are also wholesale turned off and the corresponding settings in Bluerock have no affect (they currently update the option key specific to this plugin and not the default WordPress Core option).

This updates the logic in the plugin to update the default WordPress Core auto-update settings for plugins and themes any time:

- a plugin or theme is installed, deleted, or updated.
- the plugin or theme auto-update setting is toggled.

When either of those two settings are toggled off, the Core option is not updated. This ensures all plugins and themes continue to receive auto-updates unless the user explicitly disables them on the plugin or theme page.

Bluerock and the installer scripts will soon be updated to also sync the default WordPress Core options to ensure auto-updates are appropriately applied even when the Bluehost plugin is disabled or removed.

* Some lock file updates and changes from running `build`.

* Add an upgrade routine for plugin and theme auto-update settings.

* Fix typo in option name.

* Ensure `get_plugins()` is available during the 2.7.0 upgrade routine.

* add defensive get_plugins check in bh_sync_plugin_update_settings just in case

* add defensive check for wp_get_themes in bh_sync_theme_update_settings

* phpcs fixes

Co-authored-by: Dave Ryan <dave@bluehost.com>